+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 - **New**: Nullable bindings are now allowed! See the [nullability docs](https://zacsweers.github.io/metro/bindings#nullability) for more info.
 - **Enhancement**: Add diagnostics for multibindings with star projections.
 - **Enhancement**: Add diagnostic for map multibindings with nullable keys.
+- **Fix**: Fix rank processing error when the outranked binding is contributed using Metro's ContributesBinding annotation.
 - **Fix**: Fix duplicate field accessors generated for graph supertypes.
 - Add [compose navigation sample](https://github.com/ZacSweers/metro/tree/main/samples/compose-navigation-app).